en/requirements.txt
python-osc
SpeechRecognition
pyaudio
keyboard
azure-cognitiveservices-speech